refactor(FilterTypeFileTypes): extract sort logic to useContentTypeSort composable
